backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.9.2
python-dotenv==1.0.1
email-validator==2.2.0
jinja2==3.1.6
weasyprint==68.0
argon2-cffi==23.1.0
cryptography==42.0.8
